work on validate()
