+++ lib/index.js
@@ -1,3 +1,11 @@
 'use strict'
 
+var common = require('common')
+
+function validate (barcode) {
+
+}
+
+exports.validate = validate
+
 exports.upce = require('./upce')
