+++ lib/upce.js
@@ -1,6 +1,8 @@
 'use strict'
 
-var padStart = require('lodash.padstart')
+const common = require('./common')
+
+const generateCheckDigit = common.generateCheckDigit
 
 exports.expand = function (barcode) {
   if (typeof barcode !== 'string') throw new Error('Barcode must be a string')
@@ -90,25 +92,6 @@ function setCheckDigit (barcode) {
     (barcode.length !== 12 && barcode.length !== 8)
     ? barcode + generateCheckDigit(barcode)
     : barcode
-  )
-}
-
-function generateCheckDigit (barcode) {
-  return String(
-    (
-      10 - (
-        (
-          padStart(barcode, 13, '0')
-          .split('')
-          .map(function (num, idx) {
-            return (+num) * ((idx % 2 === 0) ? 3 : 1)
-          })
-          .reduce(function (prev, cur) {
-            return prev + cur
-          })
-        ) % 10
-      )
-    ) % 10
   )
 }
